// src/lib/api/generated/model/appointmentDetailsDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Agendly.Api
 * OpenAPI spec version: 1.0
 */
import type { AppointmentStatus } from './appointmentStatus';
import type { AppointmentOrigin } from './appointmentOrigin';
import type { AppointmentCustomerDto } from './appointmentCustomerDto';
import type { AppointmentServiceDto } from './appointmentServiceDto';
import type { AppointmentProfessionalDto } from './appointmentProfessionalDto';

export interface AppointmentDetailsDto {
  id: string;
  startAt: string;
  endAt: string;
  status: AppointmentStatus;
  origin: AppointmentOrigin;
  priceCharged: number;
  customer: AppointmentCustomerDto;
  service: AppointmentServiceDto;
  professional: AppointmentProfessionalDto;
}